+++ src/frontend/packages/lib_video/src/components/common/VideoWizard/index.ts
@@ -1,3 +1,4 @@
 export * from './ConfigureLiveButton';
 export * from './CreateVOD';
 export * from './CreateVOD/UploadVideoForm/';
+export * from './CreateVOD/LicenseSelect';