+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add Markdown wizard in LTI context
 - standalone website:
   - List the VOD in the contents section
+  - Integrate creation VOD
 
 ### Changed